Elaborate on J, Y, CK, and EE (#35)
These changes carefully distinguish the three ways to transcribe Y: for English, Sindarin, and Quenya which all have their own conventions.

Also, this fixes the use of doubled vowels in English, where the mode for general use is expected to write every tehta individually.

Also, the J is represented with anga instead of anca in the mode for general use for English.

Also, addresses the CK cluster, which in the mode for general use in the English language is expected to double the quesse.

Also, fixes the tengwa used as a silent carrier: anna. Hitherto, this was mistakenly wilya, and remains debatable since other transcribers are content often to use a silent carrier. Hopefully, these issues settle with more attested examples.