Address backtick inconsistencies
In prior changes, we revised the tick notation to be a backtick notation
and also eliminate the phoneme input normalizer.
In the process, X became distinguishable from KS and it did not make
sense any longer for the X` notation to degerate to KS, but remained
sensible to have KS` notation to explicitly opt-out of consolidation
into the special calma S-hook.

This change corrects the documentation and also catches some ticks
previously missed.